# Conflicts:
#	app/src/main/java/yaroslavgorbach/koropapps/vocabulary/data/statistics/local/model/SttisticsTimeEntity.kt
#	app/src/main/java/yaroslavgorbach/koropapps/vocabulary/feature/common/factory/StatisticsEntityFactory.kt
#	app/src/main/java/yaroslavgorbach/koropapps/vocabulary/feature/exercise/exercises/alphabet/presentation/AlphabetViewModel.kt
#	app/src/main/java/yaroslavgorbach/koropapps/vocabulary/feature/exercise/exercises/antonymssynonyms/presentation/AntonymsSynonymsViewModel.kt
#	app/src/main/java/yaroslavgorbach/koropapps/vocabulary/feature/exercise/exercises/associations/presentation/AssociationsViewModel.kt
#	app/src/main/java/yaroslavgorbach/koropapps/vocabulary/feature/exercise/exercises/game/presentation/GameViewModel.kt
#	app/src/main/java/yaroslavgorbach/koropapps/vocabulary/feature/exercise/exercises/jar/presentation/JarViewModel.kt
#	app/src/main/java/yaroslavgorbach/koropapps/vocabulary/feature/exercise/exercises/narrator/presentation/NarratorViewModel.kt
#	app/src/main/java/yaroslavgorbach/koropapps/vocabulary/feature/exercise/exercises/rememberall/presentation/RememberAllViewModel.kt
#	app/src/main/java/yaroslavgorbach/koropapps/vocabulary/feature/exercise/exercises/tautograms/presentation/TautogramsViewModel.kt
#	app/src/main/java/yaroslavgorbach/koropapps/vocabulary/feature/exercise/exercises/ten/presentation/TenViewModel.kt

+++ app/src/main/java/yaroslavgorbach/koropapps/vocabulary/feature/common/factory/StatisticsEntityFactory.kt
@@ -15,7 +15,7 @@ class StatisticsEntityFactory {
         )
     }
 
-    fun createTimeEntity(exerciseName: ExerciseName, value: Float): StatisticsTimeEntity {
+    fun createTimeEntity(exerciseName: ExerciseName, value: Long): StatisticsTimeEntity {
         return StatisticsTimeEntity(
             exerciseNameRes = exerciseName.id,
             value = value,
